shop_app/requirements.txt
# web framework
Django==4.1.6
djangorestframework==3.14.0
gunicorn==20.1.0
uvicorn==0.20.0
asyncio==3.4.3

# PostgreSQL database adapter
psycopg2-binary==2.9.5

# shop_domain
pydantic==1.10.5

# celery
celery==5.2.7
redis==4.5.1
